+++ src/util/html.ts
@@ -1,5 +1,5 @@
 import * as sax from 'sax';
-import { Link, Links } from '../link';
+import { Link } from '../link';
 
 export function parseHtmlLinks(contextUri: string, body: string): Link[] {